fix: sync ns-3-alibabacloud submodule files with Phase 2 changes
Cherry-picked reverted commits from submodule reflog:
- feat: decoupled replay Phase 2 (1598bbc)
- fix: GPUType enum, fct_writer format string (f0e19bb)
- refactor: inline SendFlow, remove _QPS_PER_CONNECTION_ (6843092)
- fix: sequential step numbering in main.cc (64a3613)

+++ ns-3-alibabacloud/simulation/scratch/decoupled_replay/topology_reader.h
@@ -66,7 +66,7 @@ inline uint64_t get_nic_rate(NodeContainer &n) {
 
 // Extracted from: common.h:175-179
 inline void get_pfc(FILE *fout, Ptr<QbbNetDevice> dev, uint32_t type) {
-    fprintf(fout, "%lu %u %u %u %u\n", Simulator::Now().GetTimeStep(),
+    fprintf(fout, "%lld %u %u %u %u\n", (long long)Simulator::Now().GetTimeStep(),
             dev->GetNode()->GetId(), dev->GetNode()->GetNodeType(),
             dev->GetIfIndex(), type);
 }
@@ -944,7 +944,7 @@ inline void SetupNetwork(void (*qp_finish)(FILE *, Ptr<RdmaQueuePair>),
                 maxRtt = rtt;
         }
     }
-    printf("maxRtt=%lu maxBdp=%lu\n", maxRtt, maxBdp);
+    printf("maxRtt=%llu maxBdp=%llu\n", (unsigned long long)maxRtt, (unsigned long long)maxBdp);
 
     for (uint32_t i = 0; i < node_num; i++) {
         if (n.Get(i)->GetNodeType() == 1) {
